Fix extra updates when merging two layers below a layer with a mask
On layer removal we should use no-filthy updates, not the filthy
ones. It lets us avoid the extra mask regeneration of the layers
above.

The patch may cause regressions, so the follwing cases should be tested
really well:

1) Removal of a layer (or merge-down with non-stardard blendmode)
   * just a layer
   * a layer with transform mask
   * a layer with layer style
   * deep in a hierarchy of group layers
2) Undo/Redo of such removal

After all the actions, the canvas should be updated properly, without
any traces or artifacts.

BUG:488133