Replace raster cursors with SVG
This will look nicer on high dpi displays. Each file has been marked
with license and original artist creating it. It removes the old files.

The cursor cache code is ifdeffed for android because large cursors get
glitchy rendering there.

For X11, it is also ifdeffed, so that the hotspot is scaled appropriately.